makes change

+++ services/bookservice.js
@@ -2,15 +2,19 @@ const books = require("../models/bookmodels")
 const counter = require("../models/autoinc");
 
 //shows all book in database
-const showAllBooks = async()=>{
-    const findallbooks = await books.find();
+const showAllBooks = async () => {
+    try {
+        const findallbooks = await books.find();
+    } catch (err) {
+        throw err;
+    }
     return findallbooks;
 }
 
 //enter a book in database
-const enterBooks = async(storedValue,req,res)=>{
+const enterBooks = async (storedValue, req, res) => {
 
-    try{
+    try {
         let seqId;
         const counterData = await counter.findOneAndUpdate(
             { id: "Autoval" },
@@ -43,39 +47,63 @@ const enterBooks = async(storedValue,req,res)=>{
         })
         return postbooks;
 
-    }catch(err){
+    } catch (err) {
         throw err;
     }
 }
 
-const getBook = async(id,req,res)=>{
+const getBook = async (id, req, res) => {
 
-    const getbooks = await books.findOne({bookId : id});
-    if(!getbooks){
-        res.status(400);
-        throw new Error("There's no book in this ID");
+    try {
+        const getbooks = await books.findOne({ bookId: id });
+        if (!getbooks) {
+            res.status(400);
+            throw new Error("There's no book in this ID");
+        }
+    }
+    catch (err) {
+        throw err;
     }
 
     return getbooks;
 }
 
-const putbook = async(id,req,res)=>{
-    const getbook = await books.findOne({bookId : id});
+const putbook = async (data) => {
+
+    const id = data.bookId;
+    const getbook = await books.findOne({bookId:id});
         if(!getbook){
-            res.status(400);
-            throw new Error("Wrong ID");
+            throw { status:400,message:"There's no book in this ID"}  
         }
-        const updatebook = await books.findOneAndUpdate({id:id}, req.body, {new : true});
-        return updatebook;
+
+    const newdata = {
+        bookId:id,
+        title:data.title,
+        author:data.author,
+        price:data.price,
+        number_of_books:data.number_of_books
+    }
+    const updatebook = await books.findOneAndUpdate({ bookId: id }, newdata, { new: true });
+    return updatebook;
 }
 
-const erasebook = async(id,req,res)=>{
+const erasebook = async (id, req, res) => {
 
-    const ans = await books.findOneAndDelete({bookId:id});
-    if(!ans){
-        return res.status(400).json({message:ans.message})
+    try {
+        const getbook = await books.findOne({ bookId: id });
+        if (!getbook) {
+            res.status(400);
+            throw new Error("There's no book in this ID");
+        }
+
+        const ans = await books.findOneAndDelete({ bookId: id });
+        if (!ans) {
+            return res.status(400).json({ message: ans.message })
+        }
     }
-
+    catch (err) {
+        throw err;
+    }
     return ans;
 }
